Module Sc_lib

module C_concept : sig ... end

Representation of features that are specific to the C language and its semantics.

module C_utils : sig ... end

Helpers for starting clang processes.

module Cil_utils : sig ... end

Helpers for Cil datatypes manipulation.

module Config : sig ... end

The tool configuration manager.

module Core_options : sig ... end

module Ctypes_bridge : sig ... end

An interface for representing C values of any type.

module Exceptions : sig ... end

module Ezcmd : sig ... end

Command line manager for external tools.

module Eztoml : sig ... end

A manager for manipulating toml files, used for configuring the orchestrator.

module Eztoml_helpers : sig ... end

Module centralizing helpers for defining rows in Eztoml. This is to avoid duplicating type definitions.

module Fuzzing : sig ... end

module Sharing : sig ... end

module Symbolization : sig ... end

Generic symbolization code for flattened heap and arguments.

module Tool : sig ... end

Tool manager. Allows to define new tools that will manipulate the project (Klee, Libfuzzer, CBMC, Frama-C plug-ins, ...)

module Tool_utils : sig ... end

Helpers for tool management.

module Types : sig ... end

Miscellaneous types used in the orchestrator and some helpers.

module Utils : sig ... end

Helpers for the orhcestrator datatypes.

module Version : sig ... end