Module C_concept.AccessPath

Access paths for aggregated C values (structs, unions, arrays)

type normal

type reversed

type atomic

type _ t

The representation of a C value. It can be atomic (atomic t) or aggregating values (normal/reversed t). The normal and reversed represent the same value, but the reversed version makes the append functions slightly more efficient.

val append : 'k t -> atomic t -> reversed t

append t a

Extends t with the atomic element a.

val append_f : 'k t -> ?deref:bool -> string -> reversed t

append_f t ?deref str

Extends t with a field str. If deref is set to false (false by default), it represents a simple data; otherwise, it represents a pointer to the data.

val append_a : 'k t -> int -> reversed t

append_a t i

Extends t with an array cell (the ith cell).

val append_aa : 'k t -> reversed t

append_a t i

Extends t with an array cell (any array cell).

val root : string option -> normal t

Builds a new value. (TODO: document the argument).

val root_rev : string option -> reversed t

Same as root, but its reversed representation.

val normal : 'k t -> normal t

Normalizes the value into its canonical representation.

val reverse : 'k t -> reversed t

Transforms the value into its reversed canonical representation.

val suff : normal t -> normal t

TODO: documentation.

val from_ref_root : normal t -> normal t

TODO: documentation.

val print' : ?strip_first_punct:bool -> 'k t Basics.PPrt.pp

Prints a C value. If strip_first_punct is set to true (false by default), it prints the first field as is, without the accessor character.

val print : 'k t Basics.PPrt.pp

Equivalent to print' ~strip_first_punct:false