Module Sc_lib.C_utils

Helpers for starting clang processes.

val clangxx_cmd : ('a, Stdlib.Format.formatter, unit, string) Stdlib.format4 -> 'a

Prints the format prefixed with Config.Env.clangxx_exe.

val clang_cmd : ('a, Stdlib.Format.formatter, unit, string) Stdlib.format4 -> 'a

Prints the format prefixed with Config.Env.clang_exe.

val cpp_cmd : ('a, Stdlib.Format.formatter, unit, string) Stdlib.format4 -> 'a

Same than clang_cmd.

val clang_c : ?clang_cmd:string -> ?cflags:string list -> ?cppflags:string list -> ?o_suff:string -> [> `C ] Sc_sys.File.TYPES.file -> [> `O ] Sc_sys.File.TYPES.file Lwt.t

clang_c ?clang_cmd ?cflags ?cppflags ?osuff file

Starts the compilation of a C file and return the object file in an Lwt promise. If the compilation fails raises Failure. Param clang_cmd: the name of the clang executable (default: Config.Env.clang_exe). Param cflags: the compilation flags (default: "-g"). Parem o_suff: a string added between the file name and the ".o" file extension (clang_c ~osuff:"foo" "file.c" -> "filefoo.o")

val clang_cxx : ?clangxx_cmd:string -> ?cxxflags:string list -> ?cppflags:string list -> ?o_suff:string -> [> `CXX ] Sc_sys.File.TYPES.file -> [> `O ] Sc_sys.File.TYPES.file Lwt.t

Same as clang_c, but for C++ files.

val clang_llvm : ?cppflags:string list -> ?libs:string list -> [> `C ] Sc_sys.File.TYPES.file -> [> `BC ] Sc_sys.File.TYPES.file Lwt.t

Starts the compilation of the file with LLVM.

val clang_ld : ?to_ld_log:Sc_sys.Process.redirection -> ?ld_cmd:string -> ?res_file:(string -> string) -> ?ldflags:string list -> ?libs:string list -> ?o_files:[ `O ] Sc_sys.File.TYPES.file list -> [< `C | `O ] Sc_sys.File.TYPES.file -> [> `exe ] Sc_sys.File.TYPES.file Lwt.t

Starts a linking process. TODO: document parameters.