Module Sc_lib.Cil_utils

Helpers for Cil datatypes manipulation.

Variable declarations

type var = Cil.typ * string * Cil.attributes

type vars = var list

val var_name : var -> string

val var_names : vars -> string list

val var_names' : vars -> string list

val varset : vars -> Basics.Strings.t

varset vars is the set of all variables names in vars

val pp_vars : vars Basics.PPrt.pp

val pp_formal_decls : vars Basics.PPrt.pp

Attributes

type ptr_attrs = [

| `Naked

| `Carr

| C_concept.MemValidity.t

]

Pointer validity attributes, extended to differentiate naked pointers from cstrings (0-terminated bytes) from arrays.

This is supposed to be an attribute attached to formal arguments and structure fields, not to types.

val add_iscarray_attr : ?sz:[ `Known of int | `Var of string ] -> var -> var

val add_iscstring_attr : ?sz:[ `Known of int | `Var of string ] -> var -> var

val pp_ptr_attrs : ptr_attrs Basics.PPrt.pp

Pretty printting of pointer validity

val ptr_attrs : Cil.attributes -> ptr_attrs option

Retrives pointer attributes from a list of attributes

A type for array type attributes

type array_attributes = {

array_len : [ `Known of int | `Var of string ] option;

static : bool;

}

Specific array attributes are its length and its static/dynamic characteristic.

val get_arr_attributes : Cil.attributes -> array_attributes

Returns details on the attributes of a pointer type.

Accessors and predicates

val fun_arg_names : string -> Cil.file -> string list

fun_arg_names f file returns the names of f's formal arguments.

val fun_arg_typs : string -> Cil.file -> Cil.typ list

fun_arg_typs f file returns the type of f's formal arguments.

val fun_arg_unrolled_typs : string -> Cil.file -> Cil.typ list

fun_arg_unrolled_typs f file same as fun_arg_typs, but unrolls types.

val fun_args : string -> Cil.file -> vars

fun_args f file returns the name and type of f's formal arguments.

val fun_unrolled_args : string -> Cil.file -> vars

fun_unrolled_args f file same as fun_args, but unrolls types.

val fun_return_typ : string -> Cil.file -> Cil.typ

fun_return_typ f file returns f's return type.

val has_function : string -> Cil.file -> bool

has_a_main_function funname file Returns true iff `file` contains a function named `funname`

val glob_types : Cil.file -> Cil.typ list

glob_types file Returns the type of the global variables

val glob_unrolled_types : Cil.file -> Cil.typ list

glob_unrolled_types file Same as glob_types but unrolls types

val glob_names : Cil.file -> string list

glob_types file Returns the name of the global variables

val glob_infos : Cil.file -> vars

glob_infos file Returns the type and the name of the global variables

val glob_unrolled_infos : Cil.file -> vars

glob_unrolled_infos file Same as glob_infos but unrolls types

val string_of_typ : ?enable_static_attr:bool -> Cil.typ -> string

string_of_typ ty Returns the string representation of a type. Size of output string must be lower than 80.

val pp_typed_assignment : (Cil.typ * string * string) Basics.PPrt.pp

pp_typed_assignment fmt (typ, dest, src) uses the formatter fmt to print an assignment from src to dest, the latter being of type typ.

val fold_typ_defs : Cil.file -> (Cil.global -> 'a -> 'a) -> 'a -> 'a

fold_typ_defs file f acc folds f over every type definition in file.

val diff_globals : Cil.file -> Cil.file -> Cil.file

diff_globals cil1 cil2 Performs a diff between cil1 and cil2 by relying on source file locations to check equality.

val multidim_array_sizes : Cil.typ -> array_attributes list

Returns the list of subarrays sizes of the array type in argument. If the type is not an array, returns

Other stuff for printing and representing functions

val pp_exp : Cil.exp Basics.PPrt.pp

Pretty prints an expression.

val pp_typ_defs : Cil.file Basics.PPrt.pp

Pretty prints the type definitions in a given Cil file.

val pp_decls : ?dpref:Basics.PPrt.ufmt -> ?dsuff:Basics.PPrt.ufmt -> vars Basics.PPrt.pp

type func_env = {

glob_vars : vars;

}

The environment of a function. For now, it only gathers the file global variables.

val func_env : ?for_func:string -> Cil.file -> func_env

Initializes the global variables of a file. For now, it ignores the for_func parameter.

val pp_func_env : func_env Basics.PPrt.pp

Pretty prints an environment.

type func_repr = {

func_name : string;

func_env : func_env;

func_rtyp : Cil.typ;

func_args : vars;

}

The simplified representation of a Cil function.

type input_kind =

| Formal

| Global

val map_func_inputs : f:(input_kind -> var -> var) -> func_repr -> func_repr

val cil_func : Cil.file -> string -> func_repr

cil_func file f

Builds the simplified representation of the function f defined in file file.

val func_inputs : func_repr -> vars

val pp_fundecl : ?enable_static_attr:bool -> func_repr Basics.PPrt.pp

Pretty prints a function representant as an extern C function declaration.

module VarMap : Stdlib.Map.S with type key = var

val parse_file : [ `C ] Sc_sys.File.TYPES.file -> Cil.file