Module Sc_lib.Config

The tool configuration manager.

Configuration as module of immutable values

module type P = sig ... end

module type Core = sig ... end

module type T = sig ... end

type config

val entrypoint : config -> string

Returns the function given on the command-line argument. TODO: These may need to just be the entry point, not necessarily the ones to label and cover.

val log_level : config -> Logs.level option

val as_module : salt:string -> tag:string -> (module Core) -> config -> (module T) Lwt.t

as_module ~salt ~tag config returns a module using the configuration value in argument and output directory properly setup if successful, an error message otherwise.

salt is a string identifier that is added to the project hash, calculated from the different tools options set in the configuration file.

In turn, tag is a user-readable string for naming project directories and/or symlinks: it should be constructed based on project options like coverage criteria.

val get_config : unit -> config

get_config () Returns the configuration of the project from the command line arguments and the configuration file, if provided. Options set on the configuration file override the command line options.

val get_config_from_config_file : [ `toml ] Sc_sys.File.TYPES.file -> config

get_config_from_config_file fname Same as get_config, but only uses the configuration file fname to build the configuration.

val save_config_file : Types.run_data -> Sc_sys.File.TYPES.dir -> Toml.Types.table -> unit

save_config_file run_data output_path Concatenates the configuration file to the config.toml file inside the project directory.

val seq_of_project_config_file : Sc_sys.File.TYPES.dir -> Toml.Types.table Stdlib.Seq.t

seq_of_project_config_file output_path Returns the sequence of table that have been saved with save_config_file.

val get_table : config -> Toml.Types.table

Returns the table set by the config file. If no config file has been set, returns the empty table.

module Env : sig ... end

The environment variables used by the SeaCoral project.