Module Sc_lib.Ctypes_bridge

An interface for representing C values of any type.

type cside

Abstract type for any C-side value

type _ ctyp

Abstract representation of C types

type struct_

Types of some C types

type union_

type array_

type 'a ctyp_consumer = {

f : 'k. 'k Ctypes.typ -> 'a;

}

Polymorphic function over internal C types

type typdecls

Type of typing declarations for the naming and definition of composed types.

val empty_typdecls : typdecls

An immutable, empty set of type declarations.

val typdecls_from_cil : Cil.file -> typdecls

typdecls_from_cil cil returns a set of typing declarations built from all global type definitions in cil.

type literal_binding =

| LBStr of string

| LBMap of literal_binding Basics.StrMap.t

| LBArr of literal_binding array

| LBLst of literal_binding list

| LBNil
(*
NULL reference
*)

| LBRef of literal_binding
(*
reference to bindings
*)

Type of bindings to C values directly from C string literals; given to VAL.assign_from_literal below.

val pp_literal_binding : literal_binding Basics.PPrt.pp

Prints raw literal bindings (for debug)

exception UnexpectedLiteral of {

binding : literal_binding;

assigned : (string * Ctypes_static.boxed_typ) option;

msg : string;

}

Exception raised in assignments with literal bindings

module type ENCODING_PARAMS = sig ... end

Parameters for binary encoding

module type ENCODING = sig ... end

High-level encoding specification. This only gathers parameters as no further details about the encoding should be exported.

module type VALUE_PRINTER = sig ... end

The type of represented values, parameterized by encoding parameters.

module Printer : sig ... end

Helpers to print C types and variable declarations

type ptr_attributes = {

pv : C_concept.MemValidity.t;

depth : int;

}

View of pointer attributes, exposed during structure traversals below

module Encoding (Params : ENCODING_PARAMS) : sig ... end

Encoding-specific low-level routines

module type VAL = sig ... end

Encapsulated typed values

module type REPR = sig ... end

Representation of typed values that is parametric in encoding parameters

module type AGGREG_TYP = sig ... end

Representation of C composed types (structure, union, arrays). These can be used to interpret any sufficiently long OCaml character string as a memory representation of a value for such a type.

module type NAMED_MAPPING = sig ... end

Representation of C structured types (structure or union). These can be built directly from a list of CIL field descriptions.

module Struct : sig ... end

Declaration and representation of C struct types.

module Union : NAMED_MAPPING with type value = cside Ctypes.union and type typ = union_ ctyp

Declaration and representation of C unions.

module BoxedArray : sig ... end

Declaration and representation of C arrays.

Other utilities for easy access to Ctypes

val has_pointers : 'k Ctypes.typ -> bool

has_pointers t is true iff the C type t is composed of at least one pointer field.

Sets and maps of C types

module CTypID : sig ... end

module CTyps : Basics.SET with type elt = CTypID.t

module CTypMap : Basics.MAP with type key = CTypID.t