Module Repr.Val

Typed values

module Encoding : ENCODING with module Params = Params

module Printer : VALUE_PRINTER

type t

val blank : typ -> t

blank typ returns a new value of type typ with all bytes set to null (zero).

val from_string : typ -> string -> t

from_string typ str interprets the string of characters str as a C structured value of type typ. Raises InvalidValueRepresentation if str is not long-enough to hold such a value.

val to_string : t -> string

to_string typ v returns the C structured value v, of type typ, as a string of bytes encoded in a string.

val read : typ -> Stdlib.in_channel -> t

read typ ic acts like from_string, but reads bytes from the input channel ic.

val write : Stdlib.out_channel -> t -> unit

write oc v writes the binary representation of the C structured value v onto oc.

val typ : t -> typ

typ v returns the type of v.

val value : t -> value

value v returns the value of v.

val print : Stdlib.Format.formatter -> t -> unit

print fmt v prints the structured value v as a C-like literal using the formatter fmt; in the presence of pointers, this does not generate parseable C code.

val as_c_literal : ?lit_heap:Printer.LiteralHeap.t as 'm -> t -> string * 'm

as_c_literal ?lit_heap v constructs a parseable C literal representation for the value v, along with a possibly empty set of object literals representing dynamci memory cells.

val assign_from_literal : typdecls -> t -> literal_binding -> unit

assign_from_literal v bindings performs a deep assignent of v w.r.t the bindings described using bindings.