Module Ctypes_bridge.Encoding

Encoding-specific low-level routines

Parameters

module Params : ENCODING_PARAMS

Signature

val pp_c_ptr_decode : (string * string * string * string) Basics.PPrt.pp

pp_c_ptr_decode fmt (v, d) prints using fmt a C expression that computes a length of array encoded by a given unsigned long variable v, at a depth encoded by an unsigned int d.

val pp_cstring_buff_length_decode : (string * string) Basics.PPrt.pp

pp_cstring_buff_length_decode fmt (v, d) prints using fmt a C expression that computes a length of C character string, including the trailing '\000', encoded by a given unsigned long variable v, at a depth encoded by an unsigned int d.

val ctyp_fold_prims : _ Ctypes.typ -> ?root_access_path:C_concept.rev_access_path -> f:(C_concept.access_path -> 'a -> 'a) ctyp_consumer -> 'a -> 'a

ctyp_fold_prims ctyp ~root_access_path ~f acc folds f over every non-pointer field defined in type ctyp.

val ctyp_fold_pointers : _ Ctypes.typ -> ?root_access_path:C_concept.rev_access_path -> ?max_depth:int -> ?unfold_implicit_arrays:bool -> deep:bool -> f:(C_concept.access_path -> ptr_attributes -> 'a -> 'a) ctyp_consumer -> 'a -> 'a

ctyp_fold_pointers ctyp ~root_access_path ?unfold_implicit_arrays ~deep ~f acc folds f over every pointer field defined in type ctyp; referenced types are also visited recursively if deep holds.