Module Ctypes_bridge.Struct

Declaration and representation of C struct types.

include NAMED_MAPPING with type value = cside Ctypes.structure and type typ = struct_ ctyp

include AGGREG_TYP with type value = cside Ctypes.structure with type typ = struct_ ctyp

type typ = struct_ ctyp

C composed type.

type value = cside Ctypes.structure

C value built for the composed type.

val size : typ -> int

size typ returns the size in bytes of the memory representation for values of the composed type typ.

val definition : Stdlib.Format.formatter -> typ -> unit

print fmt typ prints the C definition of the composed type typ on the formatter fmt.

val print : Stdlib.Format.formatter -> typ -> unit

val declare : string -> typ -> typdecls -> typdecls

declare name typ typdecls declares a typ with name name into typdecls. Any previous binding of name in typdecls is erased.

val as_ctyp : typ -> f:'a ctyp_consumer -> 'a

as_ctyp typ f applies f on the `ctypes` representation of typ.

exception InvalidValueRepresentation of typ * string option

Exception raised by Val.from_string when it is given a string that is too short to hold a value of type typ.

val from_cil_fields : ?funargs:bool -> ?typdecls:typdecls -> string -> Cil_utils.var list -> typ

from_cil_fields ~funargs ~typdecls name fields defines a composed type based on a list of named CIL fields. If given and true, funargs indicates that fixed-size array types are to be interpreted as naked pointers instead of flat arrays as in usual structure fields.

module type REPR = sig ... end

Structure-specific functions

module Repr (Params : ENCODING_PARAMS) : REPR with module Params = Params

Parametric representation of typed values