Module type Ctypes_bridge.AGGREG_TYP

Representation of C composed types (structure, union, arrays). These can be used to interpret any sufficiently long OCaml character string as a memory representation of a value for such a type.

type typ

C composed type.

type value

C value built for the composed type.

val size : typ -> int

size typ returns the size in bytes of the memory representation for values of the composed type typ.

val definition : Stdlib.Format.formatter -> typ -> unit

print fmt typ prints the C definition of the composed type typ on the formatter fmt.

val print : Stdlib.Format.formatter -> typ -> unit

val declare : string -> typ -> typdecls -> typdecls

declare name typ typdecls declares a typ with name name into typdecls. Any previous binding of name in typdecls is erased.

val as_ctyp : typ -> f:'a ctyp_consumer -> 'a

as_ctyp typ f applies f on the `ctypes` representation of typ.

exception InvalidValueRepresentation of typ * string option

Exception raised by Val.from_string when it is given a string that is too short to hold a value of type typ.

module type REPR = REPR with type typ := typ and type Val.value := value

Bound representation types

module Repr (Params : ENCODING_PARAMS) : REPR with module Params = Params

Parametric representation of typed values