Module type Ctypes_bridge.REPR

Representation of typed values that is parametric in encoding parameters

module Params : ENCODING_PARAMS

Encoding parameters for the representation

type typ

Type of representable values

val size_bounds : typ -> int * int

size_bounds typ returns a pair of the minimal structure size (i.e, assuming all pointers are NULL), and maximal size (i.e., the total memory footprint with pointers unfolded according to Encoding.max_ptr_array_length)

val fold_prims : ?root_access_path:C_concept.rev_access_path -> f:(C_concept.access_path -> 'a -> 'a) ctyp_consumer -> typ -> 'a -> 'a

fold_prims ~f typ folds f over every non-pointer field defined in type typ (cf Encoding.ctyp_fold_prims).

val fold_pointers : deep:bool -> ?root_access_path:C_concept.rev_access_path -> ?max_depth:int -> ?unfold_implicit_arrays:bool -> f:(C_concept.access_path -> ptr_attributes -> 'a -> 'a) ctyp_consumer -> typ -> 'a -> 'a

fold_pointers ~deep ?unfold_implicit_arrays ~f typ folds f over every pointer field defined in type typ (cf Encoding.ctyp_fold_pointers); referenced types are visited recursively if deep holds.

module Val : VAL with type typ := typ and module Encoding.Params = Params

Typed values