Module Sc_lib.Ezcmd

Command line manager for external tools.

module type T = sig ... end

The abstract type of command line.

val make : ?key_value_link:string -> format_key:(string -> string) -> (module T)

Given a specific key formatting, returns a module for manipulating command lines. Example: `make format_key:(Format.sprintf "-%s")` will return a module in which command lines all start with a single dash. This is the module used for `frama-c` invocations.

module Make (K : sig ... end) : T

Functorized version of make

module Std : T

Standard command lines: a signe dash for one-character keys and two dashes for other keys.