Module Ezcmd.Make

Functorized version of make

Parameters

module K : sig ... end

Signature

type t

The command. It is created though a accumulation of calls of the following primitives.

val make : string -> t

make exec Creates a command line for the executable `exec`

val raw : string -> t -> t

raw input cmd Adds `input` right after the current line.

val rawf : ('a, Stdlib.Format.formatter, unit, t -> t) Stdlib.format4 -> 'a

fmt ppf cmd appends to cmd a raw argument based on the format string ppf

val base : string -> string -> t -> t

base key binding cmd Adds the binding `key binding` to the command. The key is formatted by the `format_key` function (see `Ezcmd.make`)

val optional : string option -> string option -> t -> t

optional key binding Same as `base`, but does nothing if either `key` or `binding` is None.

val if_bool : string -> bool -> t -> t

if_bool key bool cmd Adds `key` to the command iff `bool = true`.

val key : string -> t -> t

key key cmd Same as `if_bool key true cmd`

val keyf : ('a, Stdlib.Format.formatter, unit, t -> t) Stdlib.format4 -> 'a

keyf ppf cmd formats the key according to the format string ppf

val must_value : string -> string option -> t -> t

must_value key binding cmd Same as `optional (Some key) binding cmd`

val to_cmd : t -> string * string array

to_cmd cmd Returns the command as an `Lwt_process.command`.