Module Sc_lib.Eztoml

A manager for manipulating toml files, used for configuring the orchestrator.

type 'c row

type ('a, 'c) row_def_mandatory = ('a, ('c -> 'a -> 'c) -> 'c row) Eztoml_helpers.def_mandatory

Definitions for mandatory values

Suitable for rows with "primitive" types, that come with a documentation string that must specify the default value. As of now the default value is only provided for documentation purposes.

val unit : (unit, 'c) row_def_mandatory

val bool : (bool, 'c) row_def_mandatory

val int : (int, 'c) row_def_mandatory

val float : (float, 'c) row_def_mandatory

val string : (string, 'c) row_def_mandatory

val bool_list : (bool list, 'c) row_def_mandatory

val int_list : (int list, 'c) row_def_mandatory

val float_list : (float list, 'c) row_def_mandatory

val string_list : (string list, 'c) row_def_mandatory

type ('a, 'c) row_def_optional = ('a, ('c -> 'a -> 'c) -> 'c row) Eztoml_helpers.def_optional

Definitions for optional values

Similar to row_def_mandatory, except that no default value is provided.

val bool_opt : (bool, 'c) row_def_optional

val int_opt : (int, 'c) row_def_optional

val float_opt : (float, 'c) row_def_optional

val string_opt : (string, 'c) row_def_optional

type ('a, 'c) row_def_checked = ('a, ('c -> 'a -> 'c) -> 'c row) Eztoml_helpers.checked_def_with_default

Definitions for mandatory values with some additional constraint

val strictly_positive_int : (int, 'c) row_def_checked

type 'config schema

val build : 'config row list -> 'config schema

val parse : 'config -> Toml.Types.table -> 'config schema -> 'config * int

Parses the table given the schema in argument and the initial (default) option. The first config value corresponds to the configuration the tool will use and the integer an identifier corresponding to the project related options of the config. Two confugurations will have the same identifier if they share their project related values.

val print_doc : Stdlib.Format.formatter -> _ schema -> unit

val to_cmdliner : ?prefix:string -> 'config schema -> 'config Cmdliner.Term.t -> 'config Cmdliner.Term.t

Translates an eztoml schema into a cmdliner term. Each key of the schema will be a different key ('_'s are transformed in '-'s) and prefix, if provided, prefixes each key. For example, a key "bar" prefixed with "foo" will be the command line "-foo-bar".