Module Sc_lib.Eztoml_helpers

Module centralizing helpers for defining rows in Eztoml. This is to avoid duplicating type definitions.

type ('a, 'x) def_mandatory = key:string -> doc:('a Basics.PPrt.pp -> 'a -> unit) Basics.PPrt.fmt -> ?runtime:bool -> ?check:('a -> bool) -> ?check_descr:Basics.PPrt.ufmt -> default:'a -> 'x

Documentation of each argument:

doc: documentation of the option (takes in argument the default value of the option if the option is mandatory)

runtime: sets whether the option only effects the runtime behaviors of the tool, in which case it has no impact on the project layout (false by default)

check: additional check on the argument (for example, fun x -> x > 0) for positive integers)

check_descr: optional documentation of the additional check

default: a default value of the option

type ('a, 'x) def_optional = key:string -> doc:Basics.PPrt.ufmt -> ?runtime:bool -> ?check:('a -> bool) -> ?check_descr:Basics.PPrt.ufmt -> 'x

type ('a, 'x) checked_def_with_default = key:string -> doc:('a Basics.PPrt.pp -> 'a -> unit) Basics.PPrt.fmt -> ?runtime:bool -> default:'a -> 'x