Module TestRepr.Params

Encoding parameters for the representation

val max_ptr_array_length : int

Maximal size of C arrays directly reachable via pointers.

val max_non_nil_ptr_depth : int

Maximal depth at which pointers may be non-NULL, measured in number of pointer dereferences.

val max_cstring_length : int

Maximal size of C strings, excluding any additional terminating '\000'.