Module Sharing.WithRuntimeDB

High(er)-level interface built on top of the runtime-database API.

Parameters

module Log : Logs.LOG

module DB : Tool.RuntimeDatabase

Signature

val monitor_inputs_in_dir : ?filter:('a Sc_sys.File.TYPES.file -> bool) -> ?read_new:('a Sc_sys.File.TYPES.file -> DB.raw_input option Lwt.t) -> on_new:[< `Func of DB.raw_input -> unit Lwt.t | `ShareAs of string ] -> DB.runtime_data -> Sc_sys.File.TYPES.dir -> (unit -> unit Lwt.t) Lwt.t

monitor_inputs_in_dir ~filter ~read_new ~on_new rdb dirname creates a thread that monitors file creations in directory dirname: read_new is subsequently called with the name f of any file newly created in dirname (whenever filter f = true, if given) to produce an input v.

Then, if on_new = `Func f then f is then called with the result v; DB.share_input ~toolname v is called otherwise (on_new = `ShareAs toolname).

By default, read_new is DB.read_input rdb.

The function promised must be called at most once to terminate the registration of f.

Please see limitations of Utils.monitor_dir for the case dirname does not exist upon call.

val with_bidirectional_channel : ?import_suff:string -> ?read_input:(_ Sc_sys.File.TYPES.file -> DB.raw_input option Lwt.t) -> ?write_input: [< `Func of _ Sc_sys.File.TYPES.file -> DB.raw_input -> unit Lwt.t | `Link Link ] -> toolname:string -> DB.runtime_data -> Sc_sys.File.TYPES.dir -> (unit -> 'a Lwt.t) -> 'a Lwt.t

with_bidirectional_channel ?import_suff ?read_input ?write_input ~toolname rdb dirname f calls f after having setup a bi-directional exchange protocol with the runtime database rdb.

According to this protocol, each file newly created in directory dirname and whose name does not terminate with import_suff (set to ".imported" by default), is read by using read_input (DB.read_input rdb by default), and the resulting input is exported to rdb in the name of toolname.

Reciprocally, every file shared with rdb during the execution of f and until with_bidirectional_channel terminates is placed into dirname by using write_input: the latter is `Link by default, which means a filesystem link is used; alternatively a user-defined function write may be specified with `Func write. Files imported in this way are not subject to the aforementioned export mechanism.

Please see limitations of Utils.monitor_dir for the case dirname does not exist upon call; limitations pertaining to the argument write_input also apply.

val import_all_inputs : ?import_suff:string -> ?write_input: [< `Func of _ Sc_sys.File.TYPES.file -> DB.raw_input -> unit Lwt.t | `Link Link ] -> DB.runtime_data -> Sc_sys.File.TYPES.dir -> Basics.Digests.t Lwt.t

import_all_inputs ?import_suff ?write_input rdb dirname imports every input previously (given to/shared with) rdb into the directory dirname. Inputs are imported locally using write if write_input = `Func write, or using a file-system link otherwise (i.e, write_input = `Link, the default).

The directory dirname must exist when this function is called.

val share_input' : ?on_share:(Stdlib.Digest.t -> unit Lwt.t) -> toolname:string -> DB.runtime_data -> DB.raw_input -> unit Lwt.t

Version of Tool.RuntimeDatabase.share_input that discards the returned digest.