Module Sc_lib.Symbolization

Generic symbolization code for flattened heap and arguments.

type code = {

pp_c_symbol_decls : Basics.PPrt.pu;

pp_c_symbol_assigns : Basics.PPrt.pu;

}

val flattening : (module Ctypes_bridge.Struct.REPR) -> Ctypes_bridge.Struct.typ -> Cil_utils.func_repr -> code

flattening inputs_struct func_repr constructs C code that performs definitions and calls to `_SC_FLAT` and `__SC_PTR_DECL` to test the function func_repr with inputs structure inputs_struct. The latter must properly encode the environment (global variables) and arguments for the tested function. The returned structure gives two routines that print C code:

pp_c_symbol_decls declares static memory symbols for potential pointers given to the tested function;

pp_c_symbol_assigns initializes the pointers, global variables, and one local variable for each function argument.