Module Sc_lib.Tool

Tool manager. Allows to define new tools that will manipulate the project (Klee, Libfuzzer, CBMC, Frama-C plug-ins, ...)

type project_status =

| AlreadyExists

| Created

type 'a tool_result = ('a, string) Stdlib.result Lwt.t

module type Opt = sig ... end

A single tool option.

module type TomlSpec = sig ... end

A module able to grab toml schemas. This is a signature for a functor M : ToolInit -> GenericDocumenter

val opt_module : (module TomlSpec with type opt = 'o) -> Toml.Types.table -> (module Opt with type t = 'o)

Search in the tool table the option corresponding to the toml specification.

module type ToolInit = sig ... end

Tools that can initialize their workspace.

module type GenericInitializer = sig ... end

A module able to initialize a project, i.e. preprocess files for a later use. This is a signature for a functor M : ToolInit -> GenericInitializer

module MakeGenericInitializer (Opt : Opt) (C : Config.T) (TestRepr : Ctypes_bridge.Struct.REPR) (Tool : ToolInit with type opt = Opt.t) : GenericInitializer with module TestRepr = TestRepr and type input_data = Tool.input_data and type working_data = Tool.working_data and type opt = Tool.opt

module type ProjectManager = sig ... end

A database for handling all the static data, i.e. data that is not read by the tools during the analysis.

module type RuntimeDatabase = sig ... end

A runtime database, updated by the dynamic analysis tools.

module type Database = sig ... end

The signature of a module that handles the databases of a project.

module type ToolSig = sig ... end

Tesing tools signature. Must be able to initialize their workspace and launch their tool.

module type Tester = sig ... end

A module manager thet is initialized by the next function `create_tool_accessor`. For each database `Db` creating a tool accessor, a tool accessor provides:

`get_database`, returning the database used to build it;

`get_tester`, returning the tester module for each registered tool;

`register_tool_as_tester`, building a `Tester` from a `ToolSig` and registering as a tester for `Db`.

type ('data, 'x, 'rd) tool_accessor = {

get_database : unit -> ('data, 'x, 'rd) db;

get_tester : string -> ('data, 'x, 'rd) tester option;

register_tool : ('data, 'x, 'rd) tool -> unit Lwt.t;

iter_configurables : configurable_iterator;

}

and ('d, 'x, 'r) db = (module Database with type PM.working_data = 'd and type Runtime.runtime_data = 'r and type Runtime.working_data = 'x)

and ('d, 'x, 'r) tool =

| SimpleLabelTool : (module Opt with type t = 'o) * (module ToolSig with type input_data = 'x and type label_family = [ `simple ] and type opt = 'o and type runtime_data = 'r) -> ('d, 'x, 'r) tool

| AnyLabelTool : (module Opt with type t = 'o) * (module ToolSig with type input_data = 'x and type label_family = Types.Label.any and type opt = 'o and type runtime_data = 'r) -> ('d, 'x, 'r) tool

and ('d, 'x, 'r) tester =

| SimpleLabelTester : (module Tester with type input_data = 'x and type label_family = [ `simple ] and type runtime_data = 'r) -> ('d, 'x, 'r) tester

| AnyLabelTester : (module Tester with type input_data = 'x and type label_family = Types.Label.any and type runtime_data = 'r) -> ('d, 'x, 'r) tester

and configurable = (module TomlSpec)

and configurable_iterator = (configurable -> unit) -> unit

val create_tool_accessor : (module Database with type PM.working_data = 'd and type Runtime.runtime_data = 'rd and type Runtime.working_data = 'x) -> (module Config.T) -> ('d, 'x, 'rd) tool_accessor

Creates a tool accessor from a Database

module type SIMPLE_LABEL = sig ... end

Module type for tools with simple label criterion