Module Tool.MakeGenericInitializer

Parameters

module Opt : Opt

module C : Config.T

module TestRepr : Ctypes_bridge.Struct.REPR

module Tool : ToolInit with type opt = Opt.t

Signature

include TomlSpec with type opt = Tool.opt

type opt = Tool.opt

Options for the tool

val tool_name : string

The name of the tool

val toml_schema : opt Eztoml.schema

Tool options, retrieved from project configuration

val default_opt : opt

A default option value

module TestRepr = TestRepr

type input_data = Tool.input_data

type working_data = Tool.working_data

val initialize : input_data -> (project_status * (working_data, string) Stdlib.result) Lwt.t

Initializes a project. An input project is either a raw project to prepare for other tools, or a project already prepared that a tool may want to customize.