Module type Tool.ProjectManager

A database for handling all the static data, i.e. data that is not read by the tools during the analysis.

module Tool : ToolInit

include GenericInitializer with type input_data = [ `C ] Sc_sys.File.TYPES.file list

include TomlSpec

type opt

Options for the tool

val tool_name : string

The name of the tool

val toml_schema : opt Eztoml.schema

Tool options, retrieved from project configuration

val default_opt : opt

A default option value

module TestRepr : Ctypes_bridge.Struct.REPR

type input_data = [ `C ] Sc_sys.File.TYPES.file list

type working_data

val initialize : input_data -> (project_status * (working_data, string) Stdlib.result) Lwt.t

Initializes a project. An input project is either a raw project to prepare for other tools, or a project already prepared that a tool may want to customize.

val post_process : working_data -> unit Lwt.t

Post processes the project if necessary