Module type Tool.Tester

include GenericInitializer

include TomlSpec

type opt

Options for the tool

val tool_name : string

The name of the tool

val toml_schema : opt Eztoml.schema

Tool options, retrieved from project configuration

val default_opt : opt

A default option value

module TestRepr : Ctypes_bridge.Struct.REPR

type input_data

type working_data

val initialize : input_data -> (project_status * (working_data, string) Stdlib.result) Lwt.t

Initializes a project. An input project is either a raw project to prepare for other tools, or a project already prepared that a tool may want to customize.

type label_family

The label family treated by the tool

type runtime_data

val tool_kind : [ `Preprocess | `Static | `Dynamic ]

The tool kind.

val compute : shareddb:runtime_data -> working_data -> label_family Types.Label.t list -> unit tool_result

Starts the computation. For each file of the project, the database will be opened and a process will be created. Each process returns a list of proof objective identifier (int) and a status (Types.status)

Returns the statistics of the last computation of the tool.