Module Tool_utils.LabelToolDefaults

Defines default helpers for the tools.

Parameters

module LabelToolSrcInfo : LabelToolSrcInfoSig

module RuntimeDB : Tool.RuntimeDatabase

module Config : Config.P

Signature

include sig ... end

val doc : string

A short documentation for the logs of the tool.

val module_name : string

The name of the current module.

val tool_name : string

The name of the tool.

module Log : sig ... end

module Log_lwt : sig ... end

val tool_project_dir : Sc_sys.File.TYPES.dir

The directory in which the tool will work.

type label_family = LabelToolSrcInfo.label_family

type input_data = RuntimeDB.working_data

type working_data = LabelToolSrcInfo.working_data

val output_file : ?funname:string -> input_data -> [ `C ] Sc_sys.File.TYPES.file

Returns the labelized C file name that the tool will work on.