Parameter LabelToolDefaults.RuntimeDB

include Tool.GenericInitializer

include Tool.TomlSpec

type opt

Options for the tool

val tool_name : string

The name of the tool

val toml_schema : opt Eztoml.schema

Tool options, retrieved from project configuration

val default_opt : opt

A default option value

module TestRepr : Ctypes_bridge.Struct.REPR

type input_data

type working_data

type raw_input = TestRepr.Val.t

val initialize : input_data -> Types.Label.any Types.Label.t list -> working_data Lwt.t

initialize proof_objectives file initializes the runtime database and returns one working file for each function of file that is to be tested. proof_objectives must include all objectives associated with infile, whatever their status (unknown, covered, uncoverable).

TODO: It's a map from tested functions to proof objectives that we may need here.

val print_tested_fun : working_data Basics.PPrt.pp

val typdecls : working_data -> Ctypes_bridge.typdecls

val inputs_struct : working_data -> Ctypes_bridge.Struct.typ

val label_data : working_data -> Types.label_data

val tested_fun_repr : working_data -> Cil_utils.func_repr

val tested_fun : working_data -> string

val cil : working_data -> Cil.file

type runtime_data

val start : working_data -> runtime_data Lwt.t

start file returns a database of inputs and a share-able memory array for measuring coverage of the function/file identified by file.

TODO: this may deserve an option to force nullifying the db?

val input_data_file : runtime_data -> input_data

input_data_file rdb provides the input file used to construct rdb (via for_proof_objectives.

val read_input : runtime_data -> 'a Sc_sys.File.TYPES.file -> raw_input option Lwt.t

read_input rdb f reads an input value from file f; returns None if f does not contain a valid input.

val write_input : 'a Sc_sys.File.TYPES.file -> raw_input -> unit Lwt.t

write_input f v promises to write the input value v into the file named f.

val existing_input_ids : ?exclude:Basics.Digests.t -> runtime_data -> Basics.Digests.t Lwt.t

existing_input_ids ~exclude rdb is a stream of all identifiers of inputs previously registered into rdb. Members of exclude, if given, are ignored.

val existing_inputs : ?exclude:Basics.Digests.t -> runtime_data -> raw_input Types.result_view Lwt_stream.t

existing_inputs ~exclude rdb is a stream of all inputs previously registered into rdb. Inputs whose digest belongs to exclude, if given, are ignored.

val share_input : ?on_share:(Stdlib.Digest.t -> unit Lwt.t) -> toolname:string -> runtime_data -> raw_input -> Stdlib.Digest.t Lwt.t

share_input ~on_share ~toolname rdb v promises to share a given input with value v, generated by a tool named toolname, in rdb. If provided, on_share is called immediately with the identifier computed from v.

val share_status : toolname:string -> runtime_data -> [ `Cov | `Uncov | `Unk ] -> Basics.Ints.t -> unit Lwt.t

share_status ~toolname rdb status label_ids shares a set of status facts for the set of labels identified in label_ids. Those are registered as reported by the tool named toolname. May raise Invalid_argument if an identifier in label_ids does not correspond to a valid label, or Exceptions.ProofInconsistency if at least one new status contradicts already known results.

val on_new_input : runtime_data -> (raw_input Types.result_view -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t

on_new_input rdb f promises to call f with the identifier and value of any new input that is given to rdb (e.g, via share_input). The promised/returned function must be called at most once to terminate the registration of f.

val on_termination : runtime_data -> h:([> `Done | `Stop ] -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t

on_termination rdb ~h registers a handler h that is called with argument `Done when all the objectives have been satisfied, or else with `Stop when the whole process stops. on_termination promises (returns) a function that can be used to cancel the registration.

val stop_receiving_inputs : runtime_data -> unit Lwt.t

stop_receiving_inputs rdb returns a promise that is fulfilled whenever every input previously shared in rdb has been received and registered.

TODO: specify what happens if share_input is called afterwards.

val for_compiled_subprocess : ?ld_preload_var:string -> runtime_data -> Types.compiled_subprocess_info Lwt.t

for_compiled_subprocess rdb promises runtime information required for spanwed processed to access the shared coverage database.

val register_one_bypassed_input : runtime_data -> unit

register_bypassed_input rdb registers an additional test input whose coverage has been accounted for in the status store, but that is not inserted into the database.

val covinfo : runtime_data -> Types.covinfo Lwt.t

covinfo rdb returns some statistics about coverage as measured via the memory array shared to spawned processes.

val generate_test_files : exclude:Basics.Digests.t -> runtime_data -> unit Lwt.t

Generates the test files corresponding to the registered database tests.