Module type Tool_utils.LabelToolSrcInfoSig

The signature of a tool that will work on some data (often label & their status).

include ToolSrcInfoSig

val doc : string

A short documentation for the logs of the tool.

val module_name : string

The name of the current module.

val tool_name : string

The name of the tool.

type working_data

type label_family