Module Types.Label

type simple_view = {

sl_id : int;
(*
The unique ID of the proof objective
*)

sl_status : status;
(*
The current status of the proof objective
*)

sl_orig_loc : location;
(*
The location of the proof objective in the original file
*)

sl_loc : location;
(*
The location in the labelized file
*)

sl_emitter : string option;
(*
The tool that emited the status
*)

sl_exec : float;
(*
Exec time needed to be covered/proven uncoverable
*)

sl_tag : string;
(*
The label tag
*)

}

A simple label is identified by its tag (usually the coverage criterion) and its location on the source file.

type hyper_view

type 'family t

type any = [

| `simple

| `hyper

]

val get_id : [< any ] t -> int

Returns a unique identifier for the label in argument.

val get_status : [< any ] t -> status

Returns the status of the label.

val get_emitter : [< any ] t -> string option

Returns the name of the tool who emited a status.

val view_simple : [ `simple ] t -> simple_view

Returns the view of a simple label.

val create_simple : simple_view -> [ `simple ] t

From a simple label view, creates a label.

val cast : 'a t -> any t

Casts a label into its most generic type (no memory cost).

val is_covered : [< any ] t -> bool

Returns true if the label status is 'Covered _'

val is_uncoverable : [< any ] t -> bool

Returns true if the label status is 'Uncoverable'

val is_unknown : [< any ] t -> bool

Returns true if the label status is 'Unknwon'

val max_label_id : any t list -> int

Checks that the labels ids contiguous and start at 1, then returns the max label id (which must be the total number of label). Raises Invalid_arg if it does not satisfy the specification.