Module Sc_lib.Utils

Helpers for the orhcestrator datatypes.

val run_data : Sc_sys.File.TYPES.dir -> Types.run_data Lwt.t

run_data dir returns an updated runtime metadata structure w.r.t the already existing project directory dir.

val loc_of_string : string -> Types.location

Returns a location given a string filename:line

module Fails : sig ... end

val list_opt_to_list : 'a option list -> 'a list

val list_last : 'a list -> 'a

Returns the last element of a list. Fails with Invalid_argument if the list is empty

val include_file_from_fmt : Stdlib.Format.formatter -> string -> unit

Includes the file in argument

val include_file : Stdlib.out_channel -> string -> unit

Same as `include_file_from_fmt`, but with an out_channel

val copy : Stdlib.in_channel -> Stdlib.out_channel -> unit

Writes the content of in_channel in out_channel

val testfile_name : run_num:int -> run_ref_time:float -> Types.result_metadata -> string

Returns the test driver file name.

module type CACHE = sig ... end

type 't sharing_mem = {

on_share : 't -> unit Lwt.t;

filter : 't -> bool Lwt.t;

}

Structure returned by setup_sharing_mem.

val setup_sharing_mem : (module CACHE with type elt = 't) -> 't sharing_mem

setup_sharing_mem (module Set) setups and returns a Set.elt sharing_mem, that is such that any call to filter with an argument not previously given to on_share returns true.

Logs reporting

val reporter : ?max_level:Logs.level -> ?time_precision:int -> Stdlib.Format.formatter -> Logs.reporter * Logs.level option Stdlib.ref

reporter ?max_level ?time_precision ppf returns both a properly setup Logs reporter, and a reference to the level below which the reporter does not report log messages.

The returned referenced level !max_level is used to filter out messages: if it is None, nothing is printed; if it is Some l, then every message whose level is above l is printed.

time_precision is used to enable and control a timestamp prefix: if it is 0 (the default) or negative, then no timesamp is prepended. Otherwise if it is p, then each message is prepended with the number of seconds elapsed since the tool process started, measured with precision p and put in bewteen square brackets.

val combine_reporters : Logs.reporter as 'r -> 'r -> 'r

Combines two reporters so log messages can be broadcast to multiple reporters.

val log_subproc : string -> (module Logs_lwt.LOG)

log_subproc procname returns a module suitable for logging outputs of sub-processes named procname.