Module Sc_lib.Types

Miscellaneous types used in the orchestrator and some helpers.

Types definitions

type label_data = {

label_file : [ `labeldb ] Sc_sys.File.TYPES.file;
(*
The csv file containing information about the label coverage/uncoverability.
*)

labelized_file : [ `C | `labelized ] Sc_sys.File.TYPES.file;
(*
The C file with the extra labels.
*)

}

The data managed by LTest.

type project_data = {

original_file : [ `C ] Sc_sys.File.TYPES.file;
(*
The C file containing the analyzed function.
*)

label_data : label_data;
(*
The LTest data.
*)

}

The data shared by the project manager. x

type run_data = {

run_num : int;
(*
Number of SeaCoral runs on the same problem
*)

run_ref_time : float;
(*
Reference time of the current run
*)

}

Data gathering info on the previous and current runs.

type status =

| Unknown

| Covered of string list

| Uncoverable

The status of a label.

val status_string : status -> string

The string representation of a status.

val print_status : Stdlib.Format.formatter -> status -> unit

Prints the string representation of a status.

type covinfo = {

covered_ids : Basics.Ints.t;
(*
The set of covered labels.
*)

uncoverable_ids : Basics.Ints.t;
(*
The set of uncoverable labels.
*)

num_ids : int;
(*
The total number of labels.
*)

num_tests_gen : int;
(*
The total number of tests generated.
*)

num_tests_imported : int;
(*
The number of tests provided by the user and used to cover labels.
*)

}

Coverage information. Each label is identified by a unique integer.

type location = {

lline : int;

lfile : string;

}

The location of a label in a file.

module Label : sig ... end

type 'a result_view = {

raw : 'a lazy_t;

link : 'x. 'x Sc_sys.File.TYPES.file -> unit Lwt.t;

metadata : result_metadata;

}

A type allowing tools to either retrieve an actual result value (via raw), or to create a filesystem link with a file containing this value (via link).

and result_metadata = {

serialnum : int;
(*
Ordinal (starting from 1) incremented with every new registration of an input into the runtime database.
*)

toolname : string;
(*
Name of the tool that first transmitted the result.
*)

creatime : float;
(*
Entry creation time (using local timezone)
*)

id : Stdlib.Digest.t;
(*
Result digest
*)

}

type compiled_subprocess_info = {

rdsp_env : string array;
(*
Bindings to environment variables.
*)

rdsp_libs : string list;
(*
Dlls required, if any.
*)

}

Printers

val print_covinfo : Stdlib.Format.formatter -> covinfo -> unit

Pretty prints the statistics.

Raw statistics printers

val print_raw_stats_header : ?sep:Basics.PPrt.ufmt -> Stdlib.Format.formatter -> unit

val print_raw_stats : ?sep:Basics.PPrt.ufmt -> covinfo Fmt.t

module Loc : sig ... end

module LocMap : Basics.MAP with type key = location

Orchestration strategy

type strategy =

| Nothing
(*
Do nothing.
*)

| Tool of string
(*
A single tool to start.
*)

| Parallel of strategy list
(*
Strategies to launch in parallel.
*)

| Sequence of strategy list
(*
Strategies to launch sequentially.
*)

A strategy is a set of tool to start in a given order.

val print_strategy : Stdlib.Format.formatter -> strategy -> unit

Pretty prints a strategy.

val normalize_strategy : strategy -> strategy

Returns a normalized representation of a strategy.